// .PRD/erp-refactor/v3/types/lead.ts
import { z } from 'zod';

export const LeadSchema = z.object({
  id: z.string(),
  owner: z.string().optional(),
  creation: z.string().optional(),
  modified: z.string().optional(),
  modified_by: z.string().optional(),
  naming_series: z.enum(['CRM-LEAD-.YYYY.-']).optional(),
  salutation: z.string().optional(),
  first_name: z.string().optional(),
  middle_name: z.string().optional(),
  last_name: z.string().optional(),
  lead_name: z.string().optional(),
  job_title: z.string().optional(),
  gender: z.string().optional(),
  lead_owner: z.string().optional().default('__user'),
  status: z.enum(['Lead', 'Open', 'Replied', 'Opportunity', 'Quotation', 'Lost Quotation', 'Interested', 'Converted', 'Do Not Contact']).default('Lead'),
  customer: z.string().optional(),
  type: z.enum(['Client', 'Channel Partner', 'Consultant']).optional(),
  request_type: z.enum(['Product Enquiry', 'Request for Information', 'Suggestions', 'Other']).optional(),
  email_id: z.string().email().optional(),
  website: z.string().optional(),
  mobile_no: z.string().optional(),
  whatsapp_no: z.string().optional(),
  phone: z.string().optional(),
  phone_ext: z.string().optional(),
  company_name: z.string().optional(),
  no_of_employees: z.enum(['1-10', '11-50', '51-200', '201-500', '501-1000', '1000+']).optional(),
  annual_revenue: z.number().optional(),
  industry: z.string().optional(),
  market_segment: z.string().optional(),
  territory: z.string().optional(),
  fax: z.string().optional(),
  address_html: z.string().optional(),
  city: z.string().optional(),
  state: z.string().optional(),
  country: z.string().optional(),
  contact_html: z.string().optional(),
  utm_source: z.string().optional(),
  utm_content: z.string().optional(),
  utm_campaign: z.string().optional(),
  utm_medium: z.string().optional(),
  qualification_status: z.enum(['Unqualified', 'In Process', 'Qualified']).optional(),
  qualified_by: z.string().optional(),
  qualified_on: z.string().optional(),
  company: z.string().optional(),
  language: z.string().optional(),
  image: z.string().optional(),
  title: z.string().optional(),
  disabled: z.boolean().optional().default(false),
  unsubscribed: z.boolean().optional().default(false),
  blog_subscriber: z.boolean().optional().default(false),
  open_activities_html: z.string().optional(),
  all_activities_html: z.string().optional(),
  notes_html: z.string().optional(),
  notes: z.array(z.unknown()).optional(),
});

export type Lead = z.infer<typeof LeadSchema>;

export const LeadInsertSchema = LeadSchema.omit({ id: true, owner: true, creation: true, modified: true, modified_by: true });
export type LeadInsert = z.infer<typeof LeadInsertSchema>;